PMM-10705 Installing operators via OLM (#316)
* OLM POC

This is a POC to install OLM and replace hardcoded parts like PXC and
PSMDB handlers.
I added some methods to the K8sclient package that only expose methods
that already exist in the kubectl package.
I would like to replace the K8sClient and remove all method that are
specific for specific operators like PXC or PSMDB. I would have a
K8sUtil library with general methods and I would try to make it possible
to handle ANY operator with our library.
Also, kubectl shouldn't be internal, it must be possible to use it from
any other package. That's why I exposed some methods like Apply or Run
in K8sClient, because otherwsie I should write OLM specific code into
K8sClient. If it is K8s Client, it should only containt K8s generic
code, not code for specific operators/operator versions.

* Install OLM and PXC & PSMDB operators via OLM

* Ran format

* WIP

* Updates test

* Added OLM install and operators install

* Added OLM types

* PMM-10704 Added support for single node PSMDB

* PMM-10704 Fixed expose conditions

* Fixed tests

* Added Approve Install Plan

* WIP

* Code cleanup

* Working POC

* Ran format

* Upgraded deps to solve CVEs

* Added depedencies

* Revert "Added depedencies"

This reverts commit 1c7074c3a2ccf8c363d601a323901becaa99f978.

* Revert "Upgraded deps to solve CVEs"

This reverts commit ccac46eddd2128524badb4a434529c2ed1c967d6.

* Upgraded deps

* WIP

* WIP To include in pmm

* Updated API deps

* Updated deps

* Solving deps

* Fixing deps

* Updated api deps

* Code formated

The code was formatted with gci and updated the tool to the latest
version and updated the makefile because the gci parameters have been
changed in the latest version

* Added subscriptions list

* Ran format & linters

* Ran tidy

* trigger

* trigger

* Fixed test cleanup

* trigger

* Fixes for linters

* Removed unused code

* Removed unused code

* Updated test

* Fix test

* Try to fix test in EKS

* Random data for tests

* Updated code to use default namespace

* Test to fix test

* Added t.Parallel()

* Formatted files

* Added t.Parallel()

* More tests

* Run tests only for empty operator env var

* Fixed test to use the namespace from the environment variable.

* Updated OLM tests

* Disabled PSMDB tests

* Fixed pipeline

* Fixed test

* Ran tidy

* Updated make

* Delete deploymnents only when not OLM

* Fixes for linters

* Fixes for linters

+++ Makefile
@@ -72,7 +72,7 @@ ci-init:                ## Initialize CI environment
 format:                           ## Format source code
 	bin/gofumpt -l -w .
 	bin/goimports -local github.com/percona-platform/dbaas-controller -l -w .
-	bin/gci write --section Standard --section Default --section "Prefix(github.com/percona-platform/dbaas-controller)" .
+	bin/gci write --skip-generated --section "standard,default,prefix(github.com/percona-platform/dbaas-controller)" .
 
 check:                            ## Run checks/linters for the whole project
 	bin/check-license
@@ -158,13 +158,13 @@ eks-cleanup-namespace:
 
 eks-delete-operators:             ## Delete Kubernetes operators from EKS. Run this before deleting the cluster to not to leave garbage.
 	# Delete the PXC operator
-	kubectl ${KUBECTL_ARGS} delete deployment percona-xtradb-cluster-operator
+	if [ "$(PERCONA_TEST_DBAAS_OPERATOR)" != "olm" ]; then kubectl ${KUBECTL_ARGS} delete deployment percona-xtradb-cluster-operator; fi
 	# Delete the PSMDB operator
-	kubectl ${KUBECTL_ARGS} delete deployment percona-server-mongodb-operator
+	if [ "$(PERCONA_TEST_DBAAS_OPERATOR)" != "olm" ]; then kubectl ${KUBECTL_ARGS} delete deployment percona-server-mongodb-operator; fi
 
 eks-delete-current-namespace:
 	NAMESPACE=$$(kubectl config view --minify --output 'jsonpath={..namespace}'); \
-	if [ "$$NAMESPACE" != "default" ]; then kubectl delete ns "$$NAMESPACE"; fi
+	if [ "$(NAMESPACE)" != "default" ]; then kubectl delete ns "$$NAMESPACE"; fi
 
 deploy-to-pmm-server: release     ## Deploy DBaaS controller to a running ${PMM_CONTAINER} container.
 	docker cp bin/dbaas-controller ${PMM_CONTAINER}:/usr/sbin/dbaas-controller
